router: move nostr-specific methods outside
Move nostr-specific methods like "add_relay" outside of the message
router and the `Channel` trait. This gives us more control, for example
to add relays with custom flags.

Also add a blanket-implementation of `Channel` for any `Arc<C: Channel>`:
this lets us wrap our channel (in this case a `RelayPool`) in an `Arc`
so that we can keep a copy of it outside of the rotuer and call methods
directly on it. Also it allows re-using the `RelayPool`, which could be
interesting for NWC maybe.